requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
psycopg[binary]>=3.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.1.0
